Moved GetObjectValue into utils/object
