+++ v3/src/utils/object/GetObjectValue.js
@@ -1,0 +1,37 @@
+//  Source object
+//  The key as a string, or an array of keys, i.e. 'banner', or 'banner.hideBanner'
+//  The default value to use if the key doesn't exist
+
+var GetObjectValue = function (source, key, defaultValue)
+{
+    if (key.indexOf('.'))
+    {
+        var keys = key.split('.');
+        var parent = source;
+        var value = defaultValue;
+
+        //  Use for loop here so we can break early
+        for (var i = 0; i < keys.length; i++)
+        {
+            if (parent.hasOwnProperty(keys[i]))
+            {
+                //  Yes it has a key property, let's carry on down
+                value = parent[keys[i]];
+
+                parent = parent[keys[i]];
+            }
+            else
+            {
+                break;
+            }
+        }
+
+        return value;
+    }
+    else
+    {
+        return (source.hasOwnProperty(key)) ? source[key] : defaultValue;
+    }
+};
+
+module.exports = GetObjectValue;
